Move demo functions back out of core
