+++ src/core/index.js
@@ -2,13 +2,15 @@ export * from './books/';
 export * from './utilities';
 export * from './getApi';
 export * from './text-handling-functions';
+
+export * from './annotation-row-check';
+export * from './annotation-table-check';
 export * from './field-text-check';
 export * from './file-text-check';
 export * from './markdown-text-check';
+export * from './manifest-text-check';
 export * from './plain-text-check';
 export * from './tn-table-row-check';
 export * from './tn-table-text-check';
-export * from './annotation-row-check';
-export * from './annotation-table-check';
 export * from './usfm-text-check';
-export * from './book-package-check';
+export * from './yaml-text-check';
