+++ styleguide.config.js
@@ -117,39 +117,6 @@ let sections = [
 		]
 	},
 	{
-		name: 'Core Door43 Book Package Check',
-		// content: 'src/core/README.md',
-		sections: [
-			{
-				name: 'Book Package check (with raw results)',
-				content: 'src/core/book-package-check.md',
-				// description: ''
-			}
-		]
-	},
-	{
-		name: 'Core Door43 Repository Check',
-		// content: 'src/core/README.md',
-		sections: [
-			{
-				name: 'Repo check (with raw results)',
-				content: 'src/core/repo-check.md',
-				// description: ''
-			}
-		]
-	},
-	{
-		name: 'Core File Contents Check',
-		// content: 'src/core/README.md',
-		sections: [
-			{
-				name: 'File contents check (with raw results)',
-				content: 'src/core/file-check.md',
-				// description: ''
-			}
-		]
-	},
-	{
 		// The difficulty with displaying core functions this way
 		//	is that they all appear and run on a SINGLE web-page.
 		name: 'Core Filetype Checking Functions',
